requirements.txt
torch
safetensors
diffusers
transformers
accelerate
Pillow
pyinstaller
